[cmake] Fix linking tests to f2py-generated libraries on macOS
On macOS f2py generates bundles that cannot be linked at compile time,
so let's just build the main extension source files again for the
tests as static libraries and link to that as well as the static
libraries that are built from all the other source files before the
extension modules are built anyway. This seems to be the most reliable
option to me as it does not necessitate any cooperation from f2py.