Switch all oper (de)alloc to subroutines, add full prealloc option
By setting the option QMC.FixedPreallocatedVecs = 1, de- and
reallocations of state vector cache arrays should mostly be avoided
with potential benefits for performance. This however requires
preallocating the cache arrays with the maximum sizes of the state
vectors, which may be detrimental for performance. In total, this may
be beneficial in cases with small superstate sizes.

If this option is not set, no significant changes should be
observable.