Compatibility with Ninja buildsystem
In contrast to GNU Make, Ninja keeps track of Fortran module dependencies and
detects that the modules MCTQMC and MMaximumEntropy are built twice (once for
the Python interface and once for the testsuite). These two can potentially
conflict and therefore Ninja outright refuses to build the second instance.

The resolution to this problem is to not compile MCTQMC and MMaximumEntropy as
part of the Python interfaces, but as part of the CTQMCLIB and MAXENTLIB targets
and then simply link the Python targets and testsuite targets against either
library.

(Also the testsuite targets to not depend on Python at all, so I removed that dependency)